disperse the sqs and twilio 'clients' into the contexts that use them
