+++ stopcovid/send_sms/enqueue_outbound_sms.py
@@ -1,7 +1,14 @@
+import hashlib
+import json
 import logging
+import os
+from collections import defaultdict
 from typing import List, Optional
 from dataclasses import dataclass
 import uuid
+
+import boto3
+
 from stopcovid.dialog.models.events import (
     DrillStarted,
     UserValidated,
@@ -116,7 +123,47 @@ def get_outbound_sms_commands(dialog_events: List[DialogEvent]) -> List[Outbound
 
 
 def enqueue_outbound_sms_commands(dialog_events: List[DialogEvent]):
-    from stopcovid.clients import sqs
-
     outbound_messages = get_outbound_sms_commands(dialog_events)
-    sqs.publish_outbound_sms_messages(outbound_messages)
+    publish_outbound_sms_messages(outbound_messages)
+
+
+def publish_outbound_sms_messages(outbound_sms_messages: List[OutboundSMS]):
+    if not outbound_sms_messages:
+        return
+
+    sqs = boto3.resource("sqs")
+
+    queue_name = f"outbound-sms-{os.getenv('STAGE')}.fifo"
+    queue = sqs.get_queue_by_name(QueueName=queue_name)
+
+    phone_number_to_messages = defaultdict(list)
+    for message in outbound_sms_messages:
+        phone_number_to_messages[message.phone_number].append(message)
+
+    entries = [
+        {
+            "Id": str(uuid.uuid4()),
+            "MessageBody": json.dumps(
+                {
+                    "phone_number": phone,
+                    "messages": [
+                        {"body": message.body, "media_url": message.media_url}
+                        for message in messages
+                    ],
+                }
+            ),
+            "MessageDeduplicationId": _get_message_deduplication_id(messages),
+            "MessageGroupId": phone,
+        }
+        for phone, messages in phone_number_to_messages.items()
+    ]
+
+    return queue.send_messages(Entries=entries)
+
+
+def _get_message_deduplication_id(messages):
+    unique_message_ids = sorted(list(set([str(message.event_id) for message in messages])))
+    combined = "-".join(unique_message_ids)
+    m = hashlib.shake_256()
+    m.update(combined.encode("utf-8"))
+    return m.hexdigest(64)  # the length of a hex digest will be up to 64 * 2 or 128
